Refactor MultiJoinNode into a separate class
shared between ReorderJoins and GroupInnerJoinsByConnectorRuleSet
